LF-8416: Add -EPIPE error print in recover
The EPIPE maybe caused by the processer cost a lot of
time in processing first frame.

The EPIPE error means overrun or underrun, when this
error happen, the device is to be recoverred, but
the recover causes the latency of AFE pipeline changed.

Add error message to let user know this issue and let
user restart application.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>