Set audio thread sleep time upon period size
The capture and playback threads of the AFE used to query
the number of available audio input buffer frames every 100us.
However, if the audio buffer is written by ALSA at a greater
period (e.g., 8ms), then it is not needed to query the avail
frames so often.

This patch is meant to compute how often the avail frames will
be queried depending on the period size used. By default, the
query time is equal to 1/4 of the duration of the period size
using the THREAD_SLEEP_DIVISOR macro. This has the impact of
reducing the CPU consumption of the AFE compared to the
previous setting.

Signed-off-by: Ruben Charles <ruben.charles_1@nxp.com>